[224] Change user request to services folder
